requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
astroid==2.12.12
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
filelock==3.8.0
fire==0.4.0
frozenlist==1.3.3
fsspec==2022.10.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
isort==5.10.1
lazy-object-proxy==1.8.0
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
mypy==0.990
mypy-extensions==0.4.3
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.3
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.15.5
pyparsing==3.0.9
pytorch-lightning==1.8.0.post1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
timm==0.6.11
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.10.2
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.10.0
